original/requirements.txt
# Auto-Retrain System Dependencies
optuna>=3.0.0
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
schedule>=1.2.0
google-auth>=2.0.0
google-colab>=1.0.0